Closes #6
There may be documented exceptions which are not declared.
Usually, undeclared exceptions are unchecked exceptions,
which aren't required to be declared.
Due to such conditions, the number of existing exceptions will be the sum of:

1) declared but not documented exceptions +
2) documented but not declared exceptions +
3) documented and declared exceptions.

Now the exceptions documentation coverage for a given method is computed
based on these 3 kinds of exceptions.